+++ src/components/Form/DateField/DateField.tsx
@@ -37,13 +37,12 @@ export const DateField = ({ disabled, error, label, name, readOnly, setValue, va
   }, [value]);
 
   return (
-    <FieldGroup>
+    <FieldGroup name={name}>
       <Label htmlFor={name}>{label}</Label>
       <Popover open={isDatePickerOpen} onOpenChange={setIsDatePickerOpen}>
         <Popover.Trigger>
           <Input
             autoComplete="off"
-            data-cy="date-input"
             data-testid="date-input"
             disabled={disabled || readOnly}
             name={name}