+++ src/components/Form/FieldGroup/FieldGroupRoot.tsx
@@ -1,5 +1,7 @@
 import * as React from 'react';
 
-export const FieldGroupRoot: React.FC<{ children: React.ReactNode }> = ({ children }) => (
-  <div className="flex flex-col gap-3 @container">{children}</div>
+export const FieldGroupRoot: React.FC<{ children: React.ReactNode; name: string }> = ({ children, name }) => (
+  <div className="flex flex-col gap-3 @container" data-field-group={name}>
+    {children}
+  </div>
 );