+++ src/hooks/useStorage/index.ts
@@ -1,0 +1,3 @@
+export * from './useLocalStorage.ts';
+export * from './useSessionStorage.ts';
+export * from './useStorage.ts';