+++ src/hooks/useStorage/useLocalStorage.test.ts
@@ -1,0 +1,16 @@
+import { renderHook } from '@testing-library/react';
+import { describe, it, vi } from 'vitest';
+
+import { mockStorage } from '@/testing/mocks';
+
+import { useLocalStorage } from './useLocalStorage';
+
+mockStorage('localStorage');
+
+vi.mock('@/utils', () => ({ isBrowser: vi.fn(() => true) }));
+
+describe('useLocalStorage()', () => {
+  it('should render', () => {
+    renderHook(() => useLocalStorage('key', 'value'));
+  });
+});