+++ src/hooks/useStorage/useStorage.ts
@@ -6,22 +6,26 @@ import { isBrowser } from '@/utils';
 import { useEventCallback } from '../useEventCallback';
 import { useEventListener } from '../useEventListener';
 
+type StorageName = 'localStorage' | 'sessionStorage';
+
+type StorageEventMap = {
+  [K in StorageName]: CustomEvent;
+};
+
 declare global {
-  // eslint-disable-next-line @typescript-eslint/consistent-type-definitions
-  interface WindowEventMap {
-    'session-storage': CustomEvent;
-  }
+  // eslint-disable-next-line @typescript-eslint/consistent-type-definitions, @typescript-eslint/no-empty-object-type
+  interface WindowEventMap extends StorageEventMap {}
 }
 
 /**
  * Represents the options for customizing the behavior of serialization and deserialization.
- * @template T - The type of the state to be stored in session storage.
+ * @template T - The type of the state to be stored in storage.
  */
-type UseSessionStorageOptions<T> = {
+type UseStorageOptions<T> = {
   /** A function to deserialize the stored value. */
   deserializer?: (value: string) => T;
   /**
-   * If `true` (default), the hook will initialize reading the session storage. In SSR, you should set it to `false`, returning the initial value initially.
+   * If `true` (default), the hook will initialize reading the storage. In SSR, you should set it to `false`, returning the initial value initially.
    * @default true
    */
   initializeWithValue?: boolean;
@@ -30,25 +34,27 @@ type UseSessionStorageOptions<T> = {
 };
 
 /**
- * Custom hook that uses session storage to persist state across page reloads.
- * @template T - The type of the state to be stored in session storage.
- * @param {string} key - The key under which the value will be stored in session storage.
- * @param {T | (() => T)} initialValue - The initial value of the state or a function that returns the initial value.
- * @param {?UseSessionStorageOptions<T>} [options] - Options for customizing the behavior of serialization and deserialization (optional).
- * @returns {[T, Dispatch<SetStateAction<T>>]} A tuple containing the stored value and a function to set the value.
+ * Custom hook that uses local or session storage to persist state across page reloads.
+ * @template T - The type of the state to be stored in storage.
+ * @param key - The key under which the value will be stored in  storage.
+ * @param initialValue - The initial value of the state or a function that returns the initial value.
+ * @param options - Options for customizing the behavior of serialization and deserialization (optional).
+ * @returns A tuple containing the stored value and a function to set the value.
  * @public
  * @example
  * ```tsx
- * const [count, setCount] = useSessionStorage('count', 0);
+ * const [count, setCount] = useStorage('count', 0);
  * // Access the `count` value and the `setCount` function to update it.
  * ```
  */
-export function useSessionStorage<T>(
+export function useStorage<T>(
   key: string,
   initialValue: (() => T) | T,
-  options: UseSessionStorageOptions<T> = {}
+  storageName: StorageName,
+  options: UseStorageOptions<T> = {}
 ): [T, Dispatch<SetStateAction<T>>] {
   const { initializeWithValue = true } = options;
+  const storage = window[storageName];
 
   const serializer = useCallback<(value: T) => string>(
     (value) => {
@@ -64,14 +70,10 @@ export function useSessionStorage<T>(
     (value) => {
       if (options.deserializer) {
         return options.deserializer(value);
-      }
-      // Support 'undefined' as a value
-      if (value === 'undefined') {
+      } else if (value === 'undefined') {
         return undefined as unknown as T;
       }
-
       const defaultValue = initialValue instanceof Function ? initialValue() : initialValue;
-
       let parsed: unknown;
       try {
         parsed = JSON.parse(value);
@@ -79,22 +81,17 @@ export function useSessionStorage<T>(
         console.error(`Error parsing JSON: ${(err as Error).message}`);
         return defaultValue;
       }
-
       return parsed as T;
     },
     [options, initialValue]
   );
 
-  // Get from session storage then
-  // parse stored json or return initialValue
   const readValue = useCallback((): T => {
     const initialValueToUse = initialValue instanceof Function ? initialValue() : initialValue;
-
-    // Prevent build error "window is undefined" but keep keep working
     if (!isBrowser()) {
       return initialValueToUse;
     }
-    const raw = window.sessionStorage.getItem(key);
+    const raw = storage.getItem(key);
     return raw ? deserializer(raw) : initialValueToUse;
   }, [initialValue, key, deserializer]);
 
@@ -102,32 +99,20 @@ export function useSessionStorage<T>(
     if (initializeWithValue) {
       return readValue();
     }
-
     return initialValue instanceof Function ? initialValue() : initialValue;
   });
 
-  // Return a wrapped version of useState's setter function that ...
-  // ... persists the new value to sessionStorage.
   const setValue: Dispatch<SetStateAction<T>> = useEventCallback((value) => {
-    // Prevent build error "window is undefined" but keeps working
     if (!isBrowser()) {
-      console.warn(`Tried setting sessionStorage key “${key}” even though environment is not a client`);
+      console.warn(`Tried setting storage key “${key}” even though environment is not a client`);
     }
-
     try {
-      // Allow value to be a function so we have the same API as useState
       const newValue = value instanceof Function ? value(readValue()) : value;
-
-      // Save to session storage
-      window.sessionStorage.setItem(key, serializer(newValue));
-
-      // Save state
+      storage.setItem(key, serializer(newValue));
       setStoredValue(newValue);
-
-      // We dispatch a custom event so every similar useSessionStorage hook is notified
-      window.dispatchEvent(new StorageEvent('session-storage', { key }));
+      window.dispatchEvent(new StorageEvent(storageName, { key }));
     } catch (error) {
-      console.warn(`Error setting sessionStorage key “${key}”:`, error);
+      console.warn(`Error setting storage key “${key}”:`, error);
     }
   });
 
@@ -148,9 +133,9 @@ export function useSessionStorage<T>(
   // this only works for other documents, not the current one
   useEventListener('storage', handleStorageChange);
 
-  // this is a custom event, triggered in writeValueToSessionStorage
-  // See: useSessionStorage()
-  useEventListener('session-storage', handleStorageChange);
+  useEventListener(storageName, handleStorageChange);
 
   return [storedValue, setValue];
 }
+
+export type { StorageName, UseStorageOptions };